requirements.txt
flask
numpy
pillow
matplotlib
